Update lib/App/RecordStream/InputStream.pm
Calling decode_json() expects the input string to be UTF8, otherwise this croaks.
By calling $obj->decode(), the parser only deals with octets in the input without
regard for their encoding (which may not be known and can be hard to determine).  This
supports arbitrary input data and leaves the problem of encoding up to
consumers.