Support -M and -m options equivalent to perl's for all ops with snippets
A very useful convenience, for example:

    recs-grep -MList::MoreUtils=any -e 'any { {{Name}} eq $_ } qw(foo bar bat)'

The symantics of -M/-m match the documented behaviour in perlrun and the
produced Perl syntax.

Getopt bundling is required for -MFoo=bar to work, otherwise an error is
thrown regarding unknown option "-MFoo".

Note that, for the time being, use of -M requires explicit use of -e/-E.