Bump release version to 3.7.5 and change package versions to just "3.7"
The previous release version, 3.7.4, is represented as 3.007004 in the
style of Perl itself.  In the 3.7.3 release, the individual package
versions (not the release version) were 3.4.  3.4.0 would have expanded
into 3.004000, which is smaller than 3.007004, but 3.4 remains 3.4 and
is larger than 3.007004.  PAUSE doesn't like the decreasing version
numbers (for good reason) and so won't index it.

By changing the package versions back to just 3.7, this new release as
3.7.5 will be indexed by PAUSE.  New release versions with the same
module versions should still be indexed properly.

In general I prefer to keep the module versions the same as the release
version, but we won't be able to unless we force reset the existing
individual module versions (3.4 but soon 3.7) in PAUSE and we upload a
new release with module versions of the style "x.y.z". This is generally
considered the nuclear option though, and may make it hard for anyone
with an earlier version from CPAN to upgrade unless they know about the
newer versions.