Use PerlIO's support for in-memory files via a scalar ref instead of piping from echo
echo is less portable than in-memory files.  On Solaris, for example,
/bin/echo does not support the -n flag and instead faithfully prints
every single argument.

I'm not sure the -n flag is actually needed — why suppress the trailing
newline? — but using an in-memory file should be more reliable than
spawning a child process and reading from a shared pipe.

Of note: ok() is not actually available within this package, so I simply
switched to die() instead of importing Test::More and futzing with
$Test::Builder::Level.