Foo::Bar exists on backpan and at least one CPAN smoker has it installed
http://www.cpantesters.org/cpan/report/0bf6c7da-bb0b-11e3-83dd-3533f684c6c1

After conversing with Neil Bowers about his smoker (which produced the
report above), he suggested using his No::Such::Module package as a
sentinel for missing packages.  It's also on backpan but was only on
CPAN for ~72 hours (the time it takes for a deletion to occur) and is
unlikely to be installed on machines.

However, it's also pretty unlikely Foo::Bar is installed on many
machines, but we happened to find one of them!  I went a step further
and added a generated portion to the name of No::Such::Module.

Ideally we'd have an official CPAN sentinel package for tests like this,
i.e. the negative equivalent of CPAN::Test::Dummy::Perl5::Exists.  It'd
be a reserved namespace in PAUSE's database so it can't spring into
existence in the future and would never have been on cpan/backpan.