Escape curly braces in regexes to avoid deprecation warnings
This deprecation warning was added in a 5.17 development release, but
removed before 5.18 due to unforeseen syntax issues when the regex
delimiters were curly braces (m{...}).  Perl 5.21.1 re-introduces this
deprecation warning now that the syntax issues are resolved.  Presumably
the deprecation warning will remain for the next stable release, 5.22,
this time.

Escaping the braces should cause no problems on earlier perls.