Switch to Dist::Milla for the release management process
Dist::Milla is a simply a nice out-of-the-box configuration of
Dist::Zilla.  Several regular Dist::Zilla plugins are used as well.

A Makefile.PL is kept checked in so that only authors making releases
need to use Dist::Milla.  Everyone else just installs and uses the
repository as usual.  The Makefile.PL (and other generated files) may be
updated at any time by running:

    milla build --no-tgz && milla clean