Restore support for section 1 and section 3 man pages
Note that this preserves the bug of two hash keys for README.pod,
leaving only recs(3) installed and not RecordStream(3).  Since all the
documentation refers to RecordStream(3), I will be fixing this in the
future during a documentation update.

I fully intend for this to get less ugly in the near future by changing
how the docs are built (and slimming down the dist.ini again).