Add a "recs" executable which dispatches to individual operations a la git
This allows for git-style invocation of subcommands such as:

    recs fromcsv --header ...

More importantly, having a single command provides an obvious
point-of-entry for fatpacking the builtin operations into a single
bundle.

Similar to git, all that's necessary to add a subcommand is having an
executable in the PATH beginning with "recs-".  Operations are
preferentially loaded from the Perl lib paths, which includes the
builtins provided by the core distribution.  This saves a re-exec of
perl when it isn't necessary and will also run core operations when the
whole dist is fatpacked into this script (and hence there aren't
seperate recs-* scripts).