Check in fatlib/ and the fatpacked recs script
We want to keep track of changes to both of them.  The local/ directory
contains arch-specific packlists and paths which must match the current
machine when fatpack runs.  For this reason, we don't keep local/
checked in.

Note that local/, fatlib/, and the fatpacked recs script are all
excluded from the CPAN dists so that they don't increase the dist size
and aren't indexed by PAUSE by mistake.