Use Module::Pluggable to find and require aggregators/clumpers/deaggregators
It's better than the previous ad-hoc method in several ways (e.g. more
resilient to unexpected files), but most importantly version 5.1 and
newer know how to play nicely with App::FatPacker's @INC entry.  This
ensures that core packages are located even when fatpacked.

Module::Pluggable itself comes with no non-core deps, which is great.
(Though it used to be core itself, it no longer is and, besides, the
version we need was never in core.)