Provide the examples and story doc as linkable perldoc and help commands
The two documents are still provided as man pages too when installing
from CPAN.  Providing the doc as classes means that local perldoc(1)
works and websites like metacpan.org and search.cpan.org will link them
correctly on the web (instead of to manpage sites which don't have the
recs manpages).  Meanwhile, providing the help commands `recs examples`
and `recs story` means that fatpack users still have local access to the
examples and story.

It's possible that the overview doc in App::RecordStream should also be
available via a help command, but users are more likely to find that
first on the web (via a CPAN site or Github README).

Generated files are updated in the next commit so that they don't muddy
the diff of this commit.