recs: Use a more reliable way to determine fatpack-status for --version
The previous check reported a false positive when other code was loaded
with App::FatPacker even if the recs executable wasn't.

For example, a fatpacked, site-specific library in ~/.recs/site/foo.pm
would cause `recs --version` installed from CPAN to incorrectly report
itself as fatpacked.

The easy solution is to make it a simple constant that's just rewritten
like other parts of the script.